move IsMultiClient to EnvDesc
