+++ oneflow/core/framework/multi_client_session_context.cpp
@@ -113,10 +113,4 @@ Maybe<void> MultiClientSessionContext::TryInit(const ConfigProto& config_proto) 
   return Maybe<void>::Ok();
 }
 
-Maybe<bool> GlobalMultiClientEnv() {
-  Maybe<bool>* is_multi_client = Global<Maybe<bool>, MultiClient>::Get();
-  CHECK_NOTNULL_OR_RETURN(is_multi_client);
-  return *is_multi_client;
-}
-
 }  // namespace oneflow
