+++ oneflow/core/job/env_desc.cpp
@@ -14,6 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 #include "oneflow/core/job/env_desc.h"
+#include "oneflow/core/job/global_for.h"
 
 namespace oneflow {
 
@@ -47,6 +48,12 @@ int64_t EnvDesc::GetMachineId(const std::string& addr) const {
   CHECK_GE(machine_id, 0);
   CHECK_LT(machine_id, machine_num);
   return machine_id;
+}
+
+Maybe<bool> GlobalMultiClientEnv() {
+  Maybe<bool>* is_multi_client = Global<Maybe<bool>, MultiClient>::Get();
+  CHECK_NOTNULL_OR_RETURN(is_multi_client);
+  return *is_multi_client;
 }
 
 }  // namespace oneflow
